Move analysis formatting functions into separate module
New module is virtool.analyses.format
